requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
feedparser==6.0.11
newspaper3k==0.2.8
lxml[html_clean]
httpx==0.27.2
anthropic==0.34.2
edge-tts==6.1.12
moviepy==1.0.3
Pillow==10.4.0
apscheduler==3.10.4
python-dotenv==1.0.1
pydantic-settings==2.5.2
websockets==13.1
datasketch==1.6.5
python-multipart==0.0.12
aiofiles==24.1.0
PyYAML==6.0.2
